ext4: set default journal ioprio to medium real time priority
MIUI-1576453

kjournald2 is the kernel thread on which most other processes depend on
for doing the filesystem portion of their IO. As such, there exists
the equivalent of a priority inversion situation, where kjournald2
would get less priority as it should.

For this reason we set to "medium real time priority", which is higher
than regular tasks, but not infinitely powerful.

Change-Id: Ice30850d01f413e9fb4cc116514d636325955b74
Signed-off-by: xiongping1 <xiongping1@xiaomi.com>